add: Utils library. moved lookup_account_outputs_ringct to Utils.
