+++ src/supernode/utils.cpp
@@ -1,0 +1,6 @@
+#include "utils.h"
+
+Utils::Utils()
+{
+
+}
