+++ src/supernode/utils.h
@@ -1,0 +1,11 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+
+class Utils
+{
+public:
+  Utils();
+};
+
+#endif // UTILS_H
